+++ core/src/main/scala/com/krux/hyperion/h3/activity/EmrActivity.scala
@@ -15,11 +15,11 @@ trait EmrActivity[A <: EmrCluster] extends PipelineActivity[A] {
   def emrActivityFields: EmrActivityFields
   def emrActivityFieldsLens: Lens[Self, EmrActivityFields]
 
-  def preStepCommands = emrActivityFieldsLens.preStepCommands
+  def preStepCommands = emrActivityFields.preStepCommands
   def withPreStepCommand(commands: HString*): Self =
     (emrActivityFieldsLens >> 'preStepCommands).modify(self)(_ ++ commands)
 
-  def postStepCommands = emrActivityFieldsLens.postStepCommands
+  def postStepCommands = emrActivityFields.postStepCommands
   def withPostStepCommand(commands: HString*): Self =
     (emrActivityFieldsLens >> 'postStepCommands).modify(self)(_ ++ commands)